Developer

+++ modules/OSSMail/roundcube/program/lib/Roundcube/rcube_cache.php
@@ -45,6 +45,7 @@ class rcube_cache
     private $cache         = array();
     private $cache_changes = array();
     private $cache_sums    = array();
+    private $max_packet    = -1;
 
 
     /**
@@ -74,7 +75,7 @@ class rcube_cache
         else {
             $this->type  = 'db';
             $this->db    = $rcube->get_dbh();
-            $this->table = $this->db->table_name('cache');
+            $this->table = $this->db->table_name('cache', true);
         }
 
         // convert ttl string to seconds
@@ -163,7 +164,7 @@ class rcube_cache
         // Remove all keys
         if ($key === null) {
             $this->cache         = array();
-            $this->cache_changed = false;
+            $this->cache_changed = true;
             $this->cache_changes = array();
             $this->cache_sums    = array();
         }
@@ -196,10 +197,10 @@ class rcube_cache
     {
         if ($this->type == 'db' && $this->db && $this->ttl) {
             $this->db->query(
-                "DELETE FROM ".$this->table.
-                " WHERE user_id = ?".
-                " AND cache_key LIKE ?".
-                " AND expires < " . $this->db->now(),
+                "DELETE FROM {$this->table}".
+                " WHERE `user_id` = ?".
+                " AND `cache_key` LIKE ?".
+                " AND `expires` < " . $this->db->now(),
                 $this->userid,
                 $this->prefix.'.%');
         }
@@ -214,7 +215,7 @@ class rcube_cache
         $rcube = rcube::get_instance();
         $db    = $rcube->get_dbh();
 
-        $db->query("DELETE FROM " . $db->table_name('cache') . " WHERE expires < " . $db->now());
+        $db->query("DELETE FROM " . $db->table_name('cache', true) . " WHERE `expires` < " . $db->now());
     }
 
 
@@ -259,7 +260,15 @@ class rcube_cache
         }
 
         if ($this->type != 'db') {
-            if ($this->type == 'memcache') {
+            $this->load_index();
+
+            // Consistency check (#1490390)
+            if (!in_array($key, $this->index)) {
+                // we always check if the key exist in the index
+                // to have data in consistent state. Keeping the index consistent
+                // is needed for keys delete operation when we delete all keys or by prefix.
+            }
+            else if ($this->type == 'memcache') {
                 $data = $this->db->get($this->ckey($key));
             }
             else if ($this->type == 'apc') {
@@ -283,13 +292,12 @@ class rcube_cache
         }
         else {
             $sql_result = $this->db->limitquery(
-                "SELECT data, cache_key".
-                " FROM " . $this->table.
-                " WHERE user_id = ?".
-                " AND cache_key = ?".
+                "SELECT `data`, `cache_key`".
+                " FROM {$this->table}".
+                " WHERE `user_id` = ? AND `cache_key` = ?".
                 // for better performance we allow more records for one key
                 // get the newer one
-                " ORDER BY created DESC",
+                " ORDER BY `created` DESC",
                 0, 1, $this->userid, $this->prefix.'.'.$key);
 
             if ($sql_arr = $this->db->fetch_assoc($sql_result)) {
@@ -319,7 +327,7 @@ class rcube_cache
      * Writes single cache record into DB.
      *
      * @param string $key  Cache key name
-     * @param mxied  $data Serialized cache data 
+     * @param mixed  $data Serialized cache data
      *
      * @param boolean True on success, False on failure
      */
@@ -329,8 +337,22 @@ class rcube_cache
             return false;
         }
 
+        // don't attempt to write too big data sets
+        if (strlen($data) > $this->max_packet_size()) {
+            trigger_error("rcube_cache: max_packet_size ($this->max_packet) exceeded for key $key. Tried to write " . strlen($data) . " bytes", E_USER_WARNING);
+            return false;
+        }
+
         if ($this->type == 'memcache' || $this->type == 'apc') {
-            return $this->add_record($this->ckey($key), $data);
+            $result = $this->add_record($this->ckey($key), $data);
+
+            // make sure index will be updated
+            if ($result && !array_key_exists($key, $this->cache_sums)) {
+                $this->cache_changed    = true;
+                $this->cache_sums[$key] = true;
+            }
+
+            return $result;
         }
 
         $key_exists = array_key_exists($key, $this->cache_sums);
@@ -339,9 +361,8 @@ class rcube_cache
         // Remove NULL rows (here we don't need to check if the record exist)
         if ($data == 'N;') {
             $this->db->query(
-                "DELETE FROM " . $this->table.
-                " WHERE user_id = ?".
-                " AND cache_key = ?",
+                "DELETE FROM {$this->table}".
+                " WHERE `user_id` = ? AND `cache_key` = ?",
                 $this->userid, $key);
 
             return true;
@@ -350,12 +371,12 @@ class rcube_cache
         // update existing cache record
         if ($key_exists) {
             $result = $this->db->query(
-                "UPDATE " . $this->table.
-                " SET created = " . $this->db->now().
-                    ", expires = " . ($this->ttl ? $this->db->now($this->ttl) : 'NULL').
-                    ", data = ?".
-                " WHERE user_id = ?".
-                " AND cache_key = ?",
+                "UPDATE {$this->table}".
+                " SET `created` = " . $this->db->now().
+                    ", `expires` = " . ($this->ttl ? $this->db->now($this->ttl) : 'NULL').
+                    ", `data` = ?".
+                " WHERE `user_id` = ?".
+                " AND `cache_key` = ?",
                 $data, $this->userid, $key);
         }
         // add new cache record
@@ -363,8 +384,8 @@ class rcube_cache
             // for better performance we allow more records for one key
             // so, no need to check if record exist (see rcube_cache::read_record())
             $result = $this->db->query(
-                "INSERT INTO " . $this->table.
-                " (created, expires, user_id, cache_key, data)".
+                "INSERT INTO {$this->table}".
+                " (`created`, `expires`, `user_id`, `cache_key`, `data`)".
                 " VALUES (" . $this->db->now() . ", " . ($this->ttl ? $this->db->now($this->ttl) : 'NULL') . ", ?, ?, ?)",
                 $this->userid, $key, $data);
         }
@@ -414,20 +435,19 @@ class rcube_cache
 
         // Remove all keys (in specified cache)
         if ($key === null) {
-            $where = " AND cache_key LIKE " . $this->db->quote($this->prefix.'.%');
+            $where = " AND `cache_key` LIKE " . $this->db->quote($this->prefix.'.%');
         }
         // Remove keys by name prefix
         else if ($prefix_mode) {
-            $where = " AND cache_key LIKE " . $this->db->quote($this->prefix.'.'.$key.'%');
+            $where = " AND `cache_key` LIKE " . $this->db->quote($this->prefix.'.'.$key.'%');
         }
         // Remove one key by name
         else {
-            $where = " AND cache_key = " . $this->db->quote($this->prefix.'.'.$key);
+            $where = " AND `cache_key` = " . $this->db->quote($this->prefix.'.'.$key);
         }
 
         $this->db->query(
-            "DELETE FROM " . $this->table.
-            " WHERE user_id = ?" . $where,
+            "DELETE FROM {$this->table} WHERE `user_id` = ?" . $where,
             $this->userid);
     }
 
@@ -435,13 +455,12 @@ class rcube_cache
     /**
      * Adds entry into memcache/apc DB.
      *
-     * @param string  $key   Cache key name
-     * @param mxied   $data  Serialized cache data
-     * @param bollean $index Enables immediate index update
+     * @param string $key  Cache key name
+     * @param mixed  $data Serialized cache data
      *
      * @param boolean True on success, False on failure
      */
-    private function add_record($key, $data, $index=false)
+    private function add_record($key, $data)
     {
         if ($this->type == 'memcache') {
             $result = $this->db->replace($key, $data, MEMCACHE_COMPRESSED, $this->ttl);
@@ -452,17 +471,6 @@ class rcube_cache
             if (apc_exists($key))
                 apc_delete($key);
             $result = apc_store($key, $data, $this->ttl);
-        }
-
-        // Update index
-        if ($index && $result) {
-            $this->load_index();
-
-            if (array_search($key, $this->index) === false) {
-                $this->index[] = $key;
-                $data = serialize($this->index);
-                $this->add_record($this->ikey(), $data);
-            }
         }
 
         return $result;
@@ -507,10 +515,15 @@ class rcube_cache
 
         // Make sure index contains new keys
         foreach ($this->cache as $key => $value) {
-            if ($value !== null) {
-                if (array_search($key, $this->index) === false) {
-                    $this->index[] = $key;
-                }
+            if ($value !== null && !in_array($key, $this->index)) {
+                $this->index[] = $key;
+            }
+        }
+
+        // new keys added using self::write()
+        foreach ($this->cache_sums as $key => $value) {
+            if ($value === true && !in_array($key, $this->index)) {
+                $this->index[] = $key;
             }
         }
 
@@ -590,5 +603,33 @@ class rcube_cache
         }
 
         return $this->packed ? @unserialize($data) : $data;
+    }
+
+    /**
+     * Determine the maximum size for cache data to be written
+     */
+    private function max_packet_size()
+    {
+        if ($this->max_packet < 0) {
+            $this->max_packet = 2097152; // default/max is 2 MB
+
+            if ($this->type == 'db') {
+                if ($value = $this->db->get_variable('max_allowed_packet', $this->max_packet)) {
+                    $this->max_packet = $value;
+                }
+                $this->max_packet -= 2000;
+            }
+            else if ($this->type == 'memcache') {
+                $stats = $this->db->getStats();
+                $remaining = $stats['limit_maxbytes'] - $stats['bytes'];
+                $this->max_packet = min($remaining / 5, $this->max_packet);
+            }
+            else if ($this->type == 'apc' && function_exists('apc_sma_info')) {
+                $stats = apc_sma_info();
+                $this->max_packet = min($stats['avail_mem'] / 5, $this->max_packet);
+            }
+        }
+
+        return $this->max_packet;
     }
 }
